Store generate_arithmetic_parameters as separate script
