+++ core/test/arithmetic_parameters.h
@@ -2,62 +2,8 @@
 // Copyright (c) 2021 Scipp contributors (https://github.com/scipp)
 // clang-format off
 /*
- * This file was automatically generated by
-
-from itertools import product
-import sys
-
-import numpy as np
-
-
-def format_number(x):
-    if np.isposinf(x):
-        return 'INFINITY'
-    if np.isneginf(x):
-        return '-INFINITY'
-    if np.isnan(x):
-        return f'{"-" if np.sign(x) == -1 else ""}NAN'
-    return f'{x}'
-
-
-def build_param(a, b):
-    # implement behaviour of numpy 1.20
-    sign = -1 if np.sign(a) == -1 or np.sign(b) == -1 else 1
-    fd = sign * np.inf if (isinstance(a, float) or isinstance(b, float)) and b == 0 and a != 0\
-        else np.floor_divide(a, b)
-    return f'Params{{{a}, {b}, {format_number(np.true_divide(a, b))},' + \
-           f' {format_number(fd)}, {format_number(np.remainder(a, b))}}}'
-
-
-def gen_values(dtype):
-    return np.r_[np.arange(3, -4, -1), np.random.uniform(-10, 10, 5)].astype(dtype)
-
-
-np.random.seed(14653503)
-with open(sys.argv[1], 'w') as outf:
-    outf.write('// SPDX-License-Identifier: GPL-3.0-or-later\n')
-    outf.write('// Copyright (c) 2021 Scipp contributors (https://github.com/scipp)\n')
-    outf.write('// clang-format off\n')
-    outf.write('/' + '*\n * This file was automatically generated by\n\n')
-    outf.write(open(__file__, 'r').read())
-    outf.write('\n *' + '/\n\n')
-
-    outf.write('#include <array>\n\n')
-    outf.write('#include <cmath>\n\n')
-
-    outf.write('namespace {\n')
-
-    name_and_dtype = (("int", int), ("float", float))
-    for (a_name, a_dtype), (b_name, b_dtype) in product(name_and_dtype, name_and_dtype):
-        outf.write('template <class Params>\n')
-        outf.write(f'constexpr inline auto division_params_{a_name}_{b_name} = std::array{{\n')
-        for a, b in product(gen_values(a_dtype), gen_values(b_dtype)):
-            outf.write(build_param(a, b) + ',\n')
-        outf.write('};\n')
-
-    outf.write('} // namespace\n')
-    outf.write('// clang-format on\n')
-
+ * This file was automatically generated
+ * DO NOT CHANGE!
  */
 
 #include <array>
